tests

+++ preprocess.py
@@ -1,13 +1,18 @@
 import os
 import uuid
 import pandas as pd
+
 from shapely import to_geojson, from_geojson
 
+import rasterio
+import FireConsts
 from FireConsts import CONNECTIVITY_CLUSTER_KM, dirextdata, dirprpdata
-from FireClustering import do_clustering
+import FireClustering
 from FireLog import logger
-from FireIO import read_VNP14IMGTDL, read_VJ114IMGTDL, read_VNP14IMGML, read_VJ114IMGML, get_reg_shp, AFP_regfilter, AFP_setampm
-from FireMain import maybe_remove_static_sources
+import FireIO
+from FireVector import cal_hull
+from typing import Literal, Optional
+from FireTypes import Region, TimeStep
 from utils import timed
 
 # INPUT_DIR = "/projects/shared-buckets/gsfc_landslides/FEDSinput/VIIRS/"
@@ -18,14 +23,16 @@ OUTPUT_DIR = dirprpdata
 
 
 @timed
-def preprocess_region(region):
+def preprocess_region(region: Region):
+    from FireMain import maybe_remove_static_sources
+
     region = maybe_remove_static_sources(region, INPUT_DIR)
     with open(f"{OUTPUT_DIR}{region[0]}.json", "w") as f:
         f.write(to_geojson(region[1], indent=2))
 
 
 @timed
-def read_region(region):
+def read_region(region: Region):
     with open(f"{OUTPUT_DIR}{region[0]}.json", "r") as f:
         shape = from_geojson(f.read())
     return (region[0], shape)
@@ -33,64 +40,68 @@ def read_region(region):
 
 @timed
 def preprocess_landcover(filename="nlcd_export_510m_simplified.tif"):
-    import rasterio
-    from rasterio.warp import calculate_default_transform, reproject, Resampling
-
-    fnmLCT = os.path.join(dirextdata, "NLCD", filename)
-    dst_crs = f'EPSG:4326'
+    fnmLCT = os.path.join(FireConsts.dirextdata, "NLCD", filename)
+    dst_crs = f"EPSG:4326"
 
     with rasterio.open(fnmLCT) as src:
-        transform, width, height = calculate_default_transform(
-            src.crs, dst_crs, src.width, src.height, *src.bounds)
+        transform, width, height = rasterio.warp.calculate_default_transform(
+            src.crs, dst_crs, src.width, src.height, *src.bounds
+        )
         kwargs = src.meta.copy()
-        kwargs.update({
-            'crs': dst_crs,
-            'transform': transform,
-            'width': width,
-            'height': height
-        })
+        kwargs.update(
+            {"crs": dst_crs, "transform": transform, "width": width, "height": height}
+        )
 
-        with rasterio.open(fnmLCT.replace(".tif", "_latlon.tif"), 'w', **kwargs) as dst:
+        with rasterio.open(fnmLCT.replace(".tif", "_latlon.tif"), "w", **kwargs) as dst:
             for i in range(1, src.count + 1):
-                reproject(
+                rasterio.warp.reproject(
                     source=rasterio.band(src, i),
                     destination=rasterio.band(dst, i),
                     src_transform=src.transform,
                     src_crs=src.crs,
                     dst_transform=transform,
                     dst_crs=dst_crs,
-                    resampling=Resampling.nearest)
+                    resampling=rasterio.warp.Resampling.nearest,
+                )
 
 
-def preprocessed_filename(t, sat, *, region=None, suffix="", preprocessed_data_dir=OUTPUT_DIR):
+def preprocessed_filename(
+    t: TimeStep,
+    sat: Literal["NOAA20", "SNPP", "VIIRS", "TESTING123"],
+    *,
+    region: Optional[Region] = None,
+    suffix="",
+    preprocessed_data_dir=OUTPUT_DIR,
+):
     return os.path.join(
         preprocessed_data_dir,
         sat,
         *([] if region is None else [region[0]]),
-        f"{t[0]}{t[1]:02}{t[2]:02}_{t[3]}{suffix}.txt"
+        f"{t[0]}{t[1]:02}{t[2]:02}_{t[3]}{suffix}.txt",
     )
 
 
 @timed
-def preprocess_NRT_file(t, sat):
+def preprocess_NRT_file(t: TimeStep, sat: Literal["NOAA20", "SNPP"]):
     logger.info(f"preprocessing NRT file for {t[0]}-{t[1]}-{t[2]}, {sat}")
-    
+
     if sat == "SNPP":
-        df = read_VNP14IMGTDL(t, input_data_dir=INPUT_DIR)
+        df = FireIO.read_VNP14IMGTDL(t, INPUT_DIR)
     elif sat == "NOAA20":
-        df = read_VJ114IMGTDL(t, input_data_dir=INPUT_DIR)
+        df = FireIO.read_VJ114IMGTDL(t, INPUT_DIR)
     else:
         raise ValueError("please set SNPP or NOAA20 for sat")
 
     # set ampm
-    df = AFP_setampm(df)
-    
+    df = FireIO.AFP_setampm(df)
+
     # add the satellite information
     df["Sat"] = sat
 
     # return selected columns
     df = df[["Lat", "Lon", "FRP", "Sat", "DT", "DS", "YYYYMMDD_HHMM", "ampm"]]
 
+    filtered_df_paths = []
     for ampm in ["AM", "PM"]:
         time_filtered_df = df.loc[df["ampm"] == ampm]
 
@@ -98,27 +109,31 @@ def preprocess_NRT_file(t, sat):
 
         # make nested path if necessary
         os.makedirs(os.path.dirname(output_filepath), exist_ok=True)
-        
-        # save active pixels at this time step (day and ampm filter)    
+
+        # save active pixels at this time step (day and ampm filter)
         time_filtered_df.to_csv(output_filepath, index=False)
+
+        filtered_df_paths.append(output_filepath)
+
+    return filtered_df_paths
 
 
 @timed
-def preprocess_monthly_file(t, sat):
+def preprocess_monthly_file(t: TimeStep, sat: Literal["NOAA20", "SNPP"]):
     """Preprocess a monthly file for a given satellite"""
     logger.info(f"preprocessing monthly file for {t[0]}-{t[1]}, {sat}")
 
     if sat == "SNPP":
-        df = read_VNP14IMGML(t, input_data_dir=INPUT_DIR)
+        df = FireIO.read_VNP14IMGML(t, INPUT_DIR)
         df = df.loc[df["Type"] == 0]  # type filtering
     elif sat == "NOAA20":
-        df = read_VJ114IMGML(t, input_data_dir=INPUT_DIR)
+        df = FireIO.read_VJ114IMGML(t, INPUT_DIR)
         df = df.loc[df["mask"] >= 7]
     else:
         raise ValueError("please set SNPP or NOAA20 for sat")
 
     # set ampm
-    df = AFP_setampm(df)
+    df = FireIO.AFP_setampm(df)
 
     # add the satellite information
     df["Sat"] = sat
@@ -129,10 +144,14 @@ def preprocess_monthly_file(t, sat):
     days = df["YYYYMMDD_HHMM"].dt.date.unique()
     for day in days:
         for ampm in ["AM", "PM"]:
-            time_filtered_df = df.loc[(df["YYYYMMDD_HHMM"].dt.date == day) & (df["ampm"] == ampm)]
+            time_filtered_df = df.loc[
+                (df["YYYYMMDD_HHMM"].dt.date == day) & (df["ampm"] == ampm)
+            ]
 
-            output_filepath = preprocessed_filename((day.year, day.month, day.day, ampm), sat)
-            
+            output_filepath = preprocessed_filename(
+                (day.year, day.month, day.day, ampm), sat
+            )
+
             # make nested path if necessary
             os.makedirs(os.path.dirname(output_filepath), exist_ok=True)
 
@@ -141,40 +160,53 @@ def preprocess_monthly_file(t, sat):
 
 
 @timed
-def read_preprocessed(t, sat, *, region=None):
-    output_filepath = preprocessed_filename(t, sat ,region=region)
+def read_preprocessed(
+    t: TimeStep,
+    sat: Literal["NOAA20", "SNPP", "VIIRS", "TESTING123"],
+    *,
+    region: Optional[Region] = None,
+):
+    output_filepath = preprocessed_filename(t, sat, region=region)
     return pd.read_csv(output_filepath)
 
 
 @timed
-def preprocess_region_t(t, sat, region):
-    logger.info(f"filtering and clustering {t[0]}-{t[1]}-{t[2]} {t[3]}, {sat}, {region[0]}")
-    if sat == "VIIRS":
+def preprocess_region_t(t: TimeStep, sensor: Literal["VIIRS", "TESTING123"], region: Region):
+    logger.info(
+        f"filtering and clustering {t[0]}-{t[1]}-{t[2]} {t[3]}, {sensor}, {region[0]}"
+    )
+    if sensor == "VIIRS":
         df = pd.concat(
             [
                 read_preprocessed(t, sat="SNPP"),
                 read_preprocessed(t, sat="NOAA20"),
-            ], ignore_index=True
+            ],
+            ignore_index=True,
         )
     else:
-        df = read_preprocessed(t, sat=sat)
+        df = read_preprocessed(t, sat=sensor)
+
+    # if regional output already exists, exit early so we don't reprocess
+    output_filepath = preprocessed_filename(t, sensor, region=region)
+    if os.path.exists(output_filepath):
+        return
 
     # do regional filtering
-    shp_Reg = get_reg_shp(region[1])
-    df = AFP_regfilter(df, shp_Reg)
+    shp_Reg = FireIO.get_reg_shp(region[1])
+    df = FireIO.AFP_regfilter(df, shp_Reg)
 
     # return selected columns
     df = df[["Lat", "Lon", "FRP", "Sat", "DT", "DS", "YYYYMMDD_HHMM", "ampm", "x", "y"]]
 
     # do preliminary clustering using new active fire locations (assign cid to each pixel)
-    df = do_clustering(df, CONNECTIVITY_CLUSTER_KM)
+    df = FireClustering.do_clustering(df, CONNECTIVITY_CLUSTER_KM)
 
     # assign a uuid to each pixel
-    df['uuid'] = [uuid.uuid4() for _ in range(len(df.index))]
+    df["uuid"] = [uuid.uuid4() for _ in range(len(df.index))]
 
-    output_filepath = preprocessed_filename(t, sat, region=region)
-    
     # make nested path if necessary
     os.makedirs(os.path.dirname(output_filepath), exist_ok=True)
 
     df.to_csv(output_filepath, index=False)
+
+    return output_filepath
